timeout: replace nix with rustix

+++ src/uu/dd/src/progress.rs
@@ -460,7 +460,10 @@ extern "C" fn sigusr1_handler(_: std::os::raw::c_int) {
 
 #[cfg(target_os = "linux")]
 pub(crate) fn install_sigusr1_handler() -> Result<(), nix::errno::Errno> {
-    uucore::signals::install_signal_handler(nix::sys::signal::Signal::SIGUSR1, sigusr1_handler)
+    uucore::signals::install_signal_handler(
+        nix::sys::signal::Signal::SIGUSR1 as libc::c_int,
+        sigusr1_handler,
+    )
 }
 
 /// Return a closure that can be used in its own thread to print progress info.
